Re-branding ! 🍒🎙️🐶
Nipper (1884–1895) was a dog from Bristol, England, who served as the model for a painting by Francis Barraud titled "His Master's Voice".

This image was the basis for the dog-and-gramophone commercial logo, one of the world's most famous used by several audio recording and associated company brands.

The iconic image of a mixed fox/bull terrier, Nipper, looking into a phonograph became an international symbol of quality and excellence for the Victor Talking Machine Company.